refactor: move Carousel.tsx to ui/ folder
